docs/sample-apps/ai-research-assistant/requirements.txt
pixeltable>=0.2.26
yfinance>=0.2.28
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.59.6
sentence-transformers>=3.3.1